backend/requirements.txt
# ── PySpark ──────────────────────────────────────────────────────────────
pyspark==3.5.1

# ── Kafka ─────────────────────────────────────────────────────────────────
kafka-python==2.0.2
confluent-kafka==2.4.0

# ── FastAPI ───────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Database ──────────────────────────────────────────────────────────────
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# ── Auth ──────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── HTTP / OSINT clients ──────────────────────────────────────────────────
httpx==0.27.0
aiohttp==3.9.5

# ── Data / ML ─────────────────────────────────────────────────────────────
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2

# ── Utilities ─────────────────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
structlog==24.2.0
rich==13.7.1

# ── Geolocation ───────────────────────────────────────────────────────────
geoip2==4.8.0
maxminddb==2.6.0

# ── Testing ───────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-cov==5.0.0
httpx==0.27.0
